FIX: Phase 0 실행 환경 수정 및 파이프라인 검증 완료
- build.gradle: flyway-core, flyway-database-postgresql 의존성 추가
  (누락으로 Flyway가 classpath에 없어 마이그레이션 미실행 버그)
- application.yml: flyway baseline-on-migrate/baseline-version 설정 추가,
  datasource URL 환경변수 바인딩으로 변경
- V1__init_schema.sql: tier_grade SMALLINT → INTEGER (JPA 엔티티 타입 불일치 수정)
- RiotApiClient: WebClient 버퍼 256KB → 16MB (Master+ 래더 응답 크기 초과 대응)
- docker-compose.yml: postgres 포트 5432 → 5433 (로컬 postgres 충돌 방지)
- AdminController: Phase 0 수동 트리거용 어드민 엔드포인트 추가

Phase 0 검증 결과: 매치 500건 수집 → 듀오 페어 620개 → 랭킹 345개 정상 생성

+++ src/main/resources/db/migration/V1__init_schema.sql
@@ -78,7 +78,7 @@ CREATE TABLE IF NOT EXISTS duo_ranking (
     adc_champion_id       INT          NOT NULL,
     support_champion_id   INT          NOT NULL,
     score                 FLOAT        NOT NULL,
-    tier_grade            SMALLINT     NOT NULL,   -- 0=S, 1=A, 2=B, 3=C, 4=D
+    tier_grade            INTEGER      NOT NULL,   -- 0=S, 1=A, 2=B, 3=C, 4=D
     -- duo_pair_stats에서 비정규화 (sort=WIN_RATE/PICK_RATE JOIN 없이 처리)
     win_rate              FLOAT        NOT NULL,
     pick_rate             FLOAT        NOT NULL,
